requirements.txt
requests
aiohttp
pyfiglet
owotext
dadjokes
pymongo
dnspython
psutil
PyNaCl
chatbotAI
topggpy
validators
motor
dblpy
twitchAPI
humanfriendly
pillow
typing
youtube_dl
python-dotenv
pytz
statcord.py
emojis
python-youtube
pygit2
simpcalci